nlp/requirements.txt
grpcio==1.60.0
grpcio-tools==1.60.0
protobuf==4.25.0
googleapis-common-protos==1.63.0
spacy==3.7.2
dateparser==1.1.8
langdetect==1.0.9
croniter==2.0.2
pytz==2023.3
pyyaml==6.0
torch==2.1.0
transformers==4.35.0
scikit-learn==1.3.0
seqeval==1.2.2
pymorphy3==2.0.0
numpy==1.24.3